Move dynamic imports to Safenet feature folder
