+++ apps/web/src/features/myAccounts/index.tsx
@@ -4,15 +4,13 @@ import { DataWidget } from '@/features/myAccounts/components/DataWidget'
 import { useAllSafesGrouped, type AllSafeItemsGrouped } from '@/features/myAccounts/hooks/useAllSafesGrouped'
 import useTrackSafesCount from '@/features/myAccounts/hooks/useTrackedSafesCount'
 import css from '@/features/myAccounts/styles.module.css'
+import SafenetBanner from '@/features/safenet/components/new-safe/SafenetBanner'
 import useWallet from '@/hooks/wallets/useWallet'
 import madProps from '@/utils/mad-props'
 import { Box, Divider, Paper } from '@mui/material'
 import classNames from 'classnames'
-import dynamic from 'next/dynamic'
 import { useState } from 'react'
 import AccountListFilters from 'src/features/myAccounts/components/AccountListFilters'
-
-const SafenetBanner = dynamic(() => import('@/features/safenet/components/new-safe/SafenetBanner'))
 
 type MyAccountsProps = {
   safes: AllSafeItemsGrouped
