+++ apps/web/src/features/safenet/components/new-safe/SafenetNameStep/index.tsx
@@ -6,6 +6,7 @@ import NoWalletConnectedWarning from '@/components/new-safe/create/NoWalletConne
 import { useSafeSetupHints } from '@/components/new-safe/create/steps/OwnerPolicyStep/useSafeSetupHints'
 import layoutCss from '@/components/new-safe/create/styles.module.css'
 import { AppRoutes } from '@/config/routes'
+import SafenetNetworkSelector from '@/features/safenet/components/SafenetNetworkSelector'
 import useChains, { useCurrentChain } from '@/hooks/useChains'
 import { useMnemonicSafeName } from '@/hooks/useMnemonicName'
 import useWallet from '@/hooks/wallets/useWallet'
@@ -19,13 +20,10 @@ import { Box, Button, Divider, Grid, InputAdornment, SvgIcon, Tooltip, Typograph
 import MUILink from '@mui/material/Link'
 import { type SafeVersion } from '@safe-global/safe-core-sdk-types'
 import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
-import dynamic from 'next/dynamic'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import { useEffect } from 'react'
 import { FormProvider, useForm, useWatch } from 'react-hook-form'
-
-const SafenetNetworkSelector = dynamic(() => import('@/features/safenet/components/SafenetNetworkSelector'))
 
 type SetNameStepForm = {
   name: string
