+++ apps/web/src/pages/settings/setup.tsx
@@ -4,16 +4,14 @@ import SettingsHeader from '@/components/settings/SettingsHeader'
 import SpendingLimits from '@/components/settings/SpendingLimits'
 import { OwnerList } from '@/components/settings/owner/OwnerList'
 import { BRAND_NAME } from '@/config/constants'
+import SafenetSettings from '@/features/safenet/components/SafenetSettings'
 import useIsSafenetEnabled from '@/features/safenet/hooks/useIsSafenetEnabled'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import InfoIcon from '@/public/images/notifications/info.svg'
 import { Grid, Paper, Skeleton, SvgIcon, Tooltip, Typography } from '@mui/material'
 import type { NextPage } from 'next'
-import dynamic from 'next/dynamic'
 import Head from 'next/head'
 import ProposersList from 'src/components/settings/ProposersList'
-
-const SafenetSettings = dynamic(() => import('@/features/safenet/components/SafenetSettings'))
 
 const Setup: NextPage = () => {
   const { safe, safeLoaded } = useSafeInfo()
